fix another warning
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10932 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ oscam-config.c
@@ -736,7 +736,7 @@ int32_t init_fakecws(void)
 			}
 			else
 			{
-				cs_log("skipping fake cw %s because of wrong length (%u != 32)!", cw_string, strlen(cw_string));
+				cs_log("skipping fake cw %s because of wrong length (%u != 32)!", cw_string, (uint32_t)strlen(cw_string));
 			}
 		}
 	}
